tron_api_python/requirements.txt
eth-account==0.9.0
trx_utils==1.0.4
ecdsa==0.18.0
requests
python-dotenv==1.0.0

